+++ Changelog.md
@@ -1,3 +1,15 @@
+### 0.9.39
+* Some more features for flanking checks. Checked when targeting a token (1 token selected - the attacker and one target targeted) or when attacking.
+  - adv only flanking actor will gain advantage and no icon added to display flanking
+  - CE Only, flanking actor will gain CE effect "Flanking" on the attacker and you can configure that however you want, adv to attack or whatever.
+  - CE + advantage, grants advantage + whatever the CE "Flanking" effect has.
+  - CE Flanked. The flanked target gets the CE "Flanked" condition, rather than the attacker gettng falanking. Checking to see if a token is flanked is done whenever the token is targeted or an attack is rolled.
+  - CE Flanked No Conga. Flanked tokens cannot contribute to flanking other tokens, meaning the flanking conga line can't form.
+* Support for new item flag, Toggle Effect, each use of the item will toggle any associated active effects/convenient effects. One use to turn on, next use turns off. This can be a vaiable alternative to passive effects where you click to enable and click again to remove. Should also simply a range of effects currnetly done as macro.execute/macro.itemMacro where the on/off cases just enable/remove effects. For active effects (as opposed to convenient effects) toggling equires DAE 0.10.01.
+* Added a players version of the GM's damage card. Can be configured separately to GM card to show/not show damage done to NPCs, and show/not show damage done to players (with the option to provide apply damage ebuttons that the players can use themselves - instead of the DM applying damage). If players try to apply damage to a token they don't own an error (non fatal) will be thrown.
+  - Showing the damage applied to NPCs will show the damage resistances of the target in the target tooltip. But since players will see the modified damage done, it's not that much extra information.
+* Added option to "apply item effects" apply the effects but do not display apply effects buttons. In case players have twitchy fingers and hit the apply effects button when they shouldn't.
+
 ### 0.9.38
 * Bugfix for checkflanking error.